requirements.txt
pySerial == 3.5
pandas == 1.3.0
protobuf == 3.17.3
tqdm == 4.62.2